+++ pytorch_hccl_tests/osu/collectives/osu_reduce_scatter.py
@@ -8,6 +8,7 @@ from pytorch_hccl_tests.commons import (
     elaspsed_time_ms,
     get_device,
     get_device_event,
+    get_nbytes_from_dtype,
     safe_rand,
     sync_device,
 )
@@ -29,13 +30,12 @@ def reducescatter(args):
     Utils.check_numprocs(world_size, rank, limit=3)
     Utils.print_header(options.benchmark, rank)
 
-    df = pd.DataFrame(columns=["size_in_bytes", "avg_latency"])
+    df = pd.DataFrame(columns=["size_in_bytes", "avg_latency_ms"])
 
     for size in Utils.message_sizes(options):
         if size > options.large_message_size:
             options.skip = options.skip_large
             options.iterations = options.iterations_large
-        iterations = list(range(options.iterations + options.skip))
 
         recvcounts = np.zeros(world_size, dtype=np.int32)
         portion = int((size / 4) // world_size)
@@ -58,7 +58,7 @@ def reducescatter(args):
         ]
 
         dist.barrier()
-        for i in iterations:
+        for i in range(options.iterations + options.skip):
             if i == options.skip:
                 start_event = get_device_event(backend)
             dist.reduce_scatter(tensor, tensor_list, dist.ReduceOp.SUM, pg, False)
@@ -67,14 +67,15 @@ def reducescatter(args):
         dist.barrier()
 
         total_time_ms = elaspsed_time_ms(backend, start_event, end_event)
-        avg_latency = Utils.avg_lat(
+        avg_latency_ms = Utils.avg_lat(
             total_time_ms, options.iterations, world_size, device
         )
 
         if rank == 0:
-            logger.info("%-10d%18.2f" % (size, avg_latency))
+            logger.info("%-10d%18.2f" % (size, avg_latency_ms))
+            size_in_bytes = int(size) * get_nbytes_from_dtype(dtype)
             df = df.append(
-                {"size_in_bytes": int(size), "avg_latency": avg_latency},
+                {"size_in_bytes": size_in_bytes, "avg_latency_ms": avg_latency_ms},
                 ignore_index=True,
             )