+++ pytorch_hccl_tests/osu/collectives/osu_scatter.py
@@ -7,6 +7,7 @@ from pytorch_hccl_tests.commons import (
     elaspsed_time_ms,
     get_device,
     get_device_event,
+    get_nbytes_from_dtype,
     safe_rand,
     sync_device,
 )
@@ -28,13 +29,12 @@ def osu_scatter(args):
     Utils.check_numprocs(world_size, rank, limit=3)
     Utils.print_header(options.benchmark, rank)
 
-    df = pd.DataFrame(columns=["size_in_bytes", "avg_latency"])
+    df = pd.DataFrame(columns=["size_in_bytes", "avg_latency_ms"])
 
     for size in Utils.message_sizes(options):
         if size > options.large_message_size:
             options.skip = options.skip_large
             options.iterations = options.iterations_large
-        iterations = list(range(options.iterations + options.skip))
 
         tensor = safe_rand(size, dtype=dtype).to(device)
         scatter_list = None
@@ -44,7 +44,7 @@ def osu_scatter(args):
             ]
 
         dist.barrier()
-        for i in iterations:
+        for i in range(options.iterations + options.skip):
             if i == options.skip:
                 start_event = get_device_event(backend)
             dist.scatter(tensor, scatter_list, 0, pg, False)
@@ -53,13 +53,14 @@ def osu_scatter(args):
         dist.barrier()
 
         total_time_ms = elaspsed_time_ms(backend, start_event, end_event)
-        avg_latency = Utils.avg_lat(
+        avg_latency_ms = Utils.avg_lat(
             total_time_ms, options.iterations, world_size, device
         )
         if rank == 0:
-            logger.info("%-10d%18.2f" % (size, avg_latency))
+            logger.info("%-10d%18.2f" % (size, avg_latency_ms))
+            size_in_bytes = int(size) * get_nbytes_from_dtype(dtype)
             df = df.append(
-                {"size_in_bytes": int(size), "avg_latency": avg_latency},
+                {"size_in_bytes": size_in_bytes, "avg_latency_ms": avg_latency_ms},
                 ignore_index=True,
             )
     # Persist result to CSV file