+++ pytorch_hccl_tests/osu/p2p/osu_bibw.py
@@ -7,6 +7,7 @@ from pytorch_hccl_tests.commons import (
     elaspsed_time_ms,
     get_device,
     get_device_event,
+    get_nbytes_from_dtype,
     safe_rand,
     sync_device,
     wait_all,
@@ -40,7 +41,6 @@ def bibw(args):
             options.skip = options.skip_large
             options.iterations = options.iterations_large
 
-        iterations = list(range(options.iterations + options.skip))
         window_sizes = list(range(window_size))
 
         s_msg = safe_rand(size, dtype=dtype).to(device)
@@ -51,7 +51,7 @@ def bibw(args):
 
         dist.barrier()
         if rank == 0:
-            for i in iterations:
+            for i in range(options.iterations + options.skip):
                 if i == options.skip:
                     start_event = get_device_event(backend)
                 for j in window_sizes:
@@ -64,7 +64,7 @@ def bibw(args):
             end_event = get_device_event(backend)
             sync_device(backend)
         elif rank == 1:
-            for i in iterations:
+            for i in range(options.iterations + options.skip):
                 for j in window_sizes:
                     recv_requests[j] = dist.irecv(r_msg, 0, pg, 100)
                 for j in window_sizes:
@@ -73,15 +73,19 @@ def bibw(args):
                 wait_all(send_requests)
 
         if rank == 0:
-            size_in_bytes = float(size / (options.iterations * window_size))
-            total_time_ms = elaspsed_time_ms(backend, start_event, end_event)
+            size_in_bytes = int(size) * get_nbytes_from_dtype(dtype)
 
-            # 'norm_size' unit is bytes and 'time_elapsed_ns' in nanoseconds, so scaling is 10^3/1024
-            scaling = 2 * (1e3 / 1024.0)
-            bw = scaling * size_in_bytes / total_time_ms
+            # Number of total_iterations
+            total_iterations = options.iterations * window_size
+
+            total_time_ms = elaspsed_time_ms(backend, start_event, end_event)
+            total_time_sec_per_iter = total_time_ms / (1000 * total_iterations)
+
+            bw = size_in_bytes / total_time_ms
             logger.info("%-10d%18.2f" % (size, bw))
             df = df.append(
-                {"size_in_bytes": int(size), "bw_mb_per_sec": bw}, ignore_index=True
+                {"size_in_bytes": int(size), "bw_mb_per_sec": total_time_sec_per_iter},
+                ignore_index=True,
             )
 
     if rank == 0: