+++ pytorch_hccl_tests/osu/p2p/osu_latency.py
@@ -7,6 +7,7 @@ from pytorch_hccl_tests.commons import (
     elaspsed_time_ms,
     get_device,
     get_device_event,
+    get_nbytes_from_dtype,
     safe_rand,
     sync_device,
 )
@@ -30,7 +31,7 @@ def latency(args):
     # Print header
     Utils.print_header(options.benchmark, rank)
 
-    df = pd.DataFrame(columns=["size_in_bytes", "avg_latency"])
+    df = pd.DataFrame(columns=["size_in_bytes", "avg_latency_ms"])
 
     for size in Utils.message_sizes(options):
         if size > options.large_message_size:
@@ -61,14 +62,16 @@ def latency(args):
 
         total_time_ms = elaspsed_time_ms(backend, start_event, end_event)
 
-        avg_latency = Utils.avg_lat(
-            total_time_ms, 2 * options.iterations, world_size, device
+        # Divide by 2 since one messsage sent and one message received
+        avg_latency_ms = (
+            Utils.avg_lat(total_time_ms, options.iterations, world_size, device) / 2
         )
 
         if rank == 0:
-            logger.info("%-10d%18.2f" % (size, avg_latency))
+            logger.info("%-10d%18.2f" % (size, avg_latency_ms))
+            size_in_bytes = int(size) * get_nbytes_from_dtype(dtype)
             df = df.append(
-                {"size_in_bytes": int(size), "avg_latency": avg_latency},
+                {"size_in_bytes": size_in_bytes, "avg_latency_ms": avg_latency_ms},
                 ignore_index=True,
             )